修复游戏内伤害不正常（为什么改成int就没问题了呢？）

+++ LiteLoader/Header/EventAPI.h
@@ -493,7 +493,7 @@ class MobHurtEvent : public EventTemplate<MobHurtEvent> {
 public:
     Mob* mMob;
     ActorDamageSource* mDamageSource;
-    float mDamage;
+    int mDamage;
 };
 
 class MobDieEvent : public EventTemplate<MobDieEvent> {
